Anton_Engine/requirements.txt
autopep8==1.4.4
requests==2.22.0
atlassian-python-api==1.15.1
pyttsx3==2.80
wikipedia==1.4.0
beautifulsoup4==4.8.2
selenium==3.141.0
SpeechRecognition==3.8.1
PyAudio==0.2.11
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
jira==2.0.0